+++ src/stub/debug.ts
@@ -1,4 +1,4 @@
 import { noop } from '../utils';
 
-const debug: (name: string) => (...message: any[]) => void = noop as any;
+const debug: (name: string) => (...message: any[]) => void = () => noop;
 export default debug;