+++ src/target/es2018/stub/async-iterator-prototype.ts
@@ -1,0 +1,2 @@
+export const AsyncIteratorPrototype: AsyncIterable<any> | undefined =
+  Object.getPrototypeOf(Object.getPrototypeOf(async function* (): AsyncIterableIterator<any> {}).prototype);